fix: compose the elements used in the sidebar menu call
